Use local registry for docker.io
Doing this can solve issues when pulling exceeded the rate limit
from docker.io. When this happens, one can push the image from
docker.io to the local registry, which can help. If all the
necessary images are placed in the local image registry, this
will greatly speed up the boot up time.

Signed-off-by: Tong Li <tli@netapp.com>